job-service/requirements.txt
Flask
flask_cors
Flask_RESTful
numpy
onnx
onnxruntime
pypmml
pillow
pandas
